Expression::inferType() and remove FuncExp::setType()
